Add emailGatewayNumber to match Verizon
Without this set, the platform default of null is used and causes the config value emailGatewaySmsToMmsTextThreshold to be ignored and all messages sent to an email address sent as MMS.

emailGatewaySmsToMmsTextThreshold definition:
Integer value: the threshold of SMS segment number when an SMS sent to Email gateway (i.e. when emailGatewayNumber is not empty) should be turned into MMS. For example, if this is "1", when SMS text has 2 segments, it needs to be sent as MMS. "-1" indicates no such conversion can happen.

Bug: b/241322895
Test: requires manual testing by carrier
Change-Id: I0cde71cd7228370f97e705d1dad997f520731a3d