Remove KEY_ALLOW_HOLD_IN_IMS_CALL_BOOL references
This carrier config is being deprecated as part of the support
auto-unhold feature. This config key is only really used by Verizon and
they have since stated that they support holding so we are deprecating
this key. We will still need to make the updates for PARIS.

Bug: 490124810
Flag: com.android.internal.telephony.flags.support_auto_unhold
Test: atest ImsCallingTest
Test: atest CtsTelecomCujTestCases
Change-Id: I38f2b8ac49cfca9e62d02b51bd8534a8965cef05